Fixed bug where AndroidAPS and ESP would stop communicating Moved connection thread code to seperate thread
